Fixed Leopard-related numpty.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@2673 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Services/User/Jabber/RosterController.m
@@ -731,7 +731,7 @@ inline static Conversation * createChatWithPerson(id self, JabberPerson* person,
 }
 
 //Can drop?
-- (NSDragOperation)outlineView:(NSOutlineView *)outlineView validateDrop:(id < NSDraggingInfo >)info proposedItem:(id)item proposedChildIndex:(NSInteger)anIndex
+- (NSDragOperation)outlineView:(NSOutlineView *)outlineView validateDrop:(id < NSDraggingInfo >)info proposedItem:(id)item proposedChildIndex:(int)anIndex
 {
 	NSPasteboard * pboard = [info draggingPasteboard];
 	if([item isKindOfClass:[JabberPerson class]])
@@ -751,7 +751,7 @@ inline static Conversation * createChatWithPerson(id self, JabberPerson* person,
 	return NSDragOperationNone;
 }
 //Drop
-- (BOOL)outlineView:(NSOutlineView *)outlineView acceptDrop:(id < NSDraggingInfo >)info item:(id)item childIndex:(NSInteger)anIndex
+- (BOOL)outlineView:(NSOutlineView *)outlineView acceptDrop:(id < NSDraggingInfo >)info item:(id)item childIndex:(int)anIndex
 {
 	NSPasteboard * pboard = [info draggingPasteboard];
 	[pboard types];
